feat(editor-state): refine tooltip stack details
