+++ packages/editor/packages/editor-state/src/features/tooltip/colors.ts
@@ -1,0 +1,45 @@
+import { getStackValueHighlightRange } from './stackAnalysisTooltip';
+import { getStackSignatureFromSourceLine } from './sourceLine';
+
+import type { SpriteLookup } from 'glugglug';
+import type { SpriteLookups, TooltipHighlightRange } from './types';
+
+export function getTooltipLineColors(
+	line: string,
+	spriteLookups: SpriteLookups | undefined,
+	highlightRanges: TooltipHighlightRange[] = []
+): Array<SpriteLookup | undefined> {
+	if (!spriteLookups || line.length === 0) {
+		return [];
+	}
+
+	const colors: Array<SpriteLookup | undefined> = new Array(line.length).fill(undefined);
+	colors[0] = spriteLookups.fontTooltipText;
+
+	for (const range of highlightRanges) {
+		colors[range.start] = spriteLookups.fontTooltipHighlight;
+
+		if (range.end < line.length) {
+			colors[range.end] = spriteLookups.fontTooltipText;
+		}
+	}
+
+	return colors;
+}
+
+export function getSelectedLineTooltipColors(
+	line: string | undefined,
+	text: string[],
+	spriteLookups: SpriteLookups | undefined
+): Array<Array<SpriteLookup | undefined>> {
+	const stackSignature = line ? getStackSignatureFromSourceLine(line) : undefined;
+
+	return text.map((tooltipLine, index) => {
+		const highlightRange =
+			index === 0 && stackSignature === tooltipLine
+				? { start: 0, end: tooltipLine.length }
+				: getStackValueHighlightRange(tooltipLine);
+
+		return getTooltipLineColors(tooltipLine, spriteLookups, highlightRange ? [highlightRange] : []);
+	});
+}
