+++ packages/editor/packages/editor-state/src/features/tooltip/constants.ts
@@ -1,0 +1,6 @@
+export const TOOLTIP_WRAP_WIDTH = 32;
+
+export const tooltipHorizontalPaddingChars = 1;
+export const maxInlineStackItemCount = 3;
+export const maxLiveMemoryAddressLength = 10;
+export const maxLiveMemoryValueLength = 11;
