+++ packages/editor/packages/editor-state/src/features/tooltip/effect.test.ts
@@ -1,17 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import createStateManager from '@8f4e/state-manager';
 
-import tooltip, {
-	TOOLTIP_WRAP_WIDTH,
-	getInstructionNameFromSourceLine,
-	getInstructionSpecFromSourceLine,
-	getMemoryDeclarationIdFromSourceLine,
-	getSelectedLineTooltipContent,
-	getSelectedLineTooltipText,
-	getStackAnalysisTooltipText,
-	getStackSignatureFromSourceLine,
-	wrapTooltipText,
-} from './effect';
+import tooltip from './effect';
 
 import { createMockCodeBlock, createMockState } from '~/pureHelpers/testingUtils/testUtils';
 
@@ -20,123 +10,6 @@ function toCharacters(text: string): number[] {
 }
 
 describe('tooltip effect', () => {
-	it('reads the instruction name from source text', () => {
-		expect(getInstructionNameFromSourceLine('  add')).toBe('add');
-		expect(getInstructionNameFromSourceLine('push 1')).toBe('push');
-	});
-
-	it('wraps tooltip text to the configured maximum length', () => {
-		expect(wrapTooltipText('one two three four', 7)).toEqual(['one two', 'three', 'four']);
-	});
-
-	it('generates stack signatures from instruction specs', () => {
-		expect(getStackSignatureFromSourceLine('add')).toBe('add (T T -- T)');
-		expect(getStackSignatureFromSourceLine('drop')).toBe('drop (T -- )');
-		expect(getStackSignatureFromSourceLine('branchIfTrue 1')).toBe('branchIfTrue (int -- )');
-		expect(getStackSignatureFromSourceLine('load')).toBe('load (ptr -- int)');
-		expect(getStackSignatureFromSourceLine('store')).toBe('store (ptr T -- )');
-		expect(getStackSignatureFromSourceLine('storeBytes 3')).toBe('storeBytes (ptr int int int -- )');
-		expect(getStackSignatureFromSourceLine('int value')).toBe('int ( -- )');
-	});
-
-	it('resolves compiler spec documentation from a source line', () => {
-		expect(getInstructionSpecFromSourceLine('add')?.docs?.shortDescription).toBe(
-			'Adds two numbers of the same type and pushes the result.'
-		);
-		expect(getInstructionSpecFromSourceLine('int value')?.docs?.shortDescription).toBe(
-			'Declares memory storage for values used by the module.'
-		);
-		expect(getInstructionSpecFromSourceLine('; add')?.docs).toBeUndefined();
-		expect(getInstructionSpecFromSourceLine('@pos 1 2')?.docs).toBeUndefined();
-	});
-
-	it('extracts named memory declaration ids from selected source lines', () => {
-		expect(getMemoryDeclarationIdFromSourceLine('int value 1')).toBe('value');
-		expect(getMemoryDeclarationIdFromSourceLine('int[] buffer 4 48 50')).toBe('buffer');
-		expect(getMemoryDeclarationIdFromSourceLine('int 1')).toBeUndefined();
-		expect(getMemoryDeclarationIdFromSourceLine('add')).toBeUndefined();
-	});
-
-	it('wraps selected line documentation', () => {
-		expect(getSelectedLineTooltipText('push 1', 24)).toEqual([
-			'push ( -- T)',
-			'Pushes a literal, memory',
-			'value, local value,',
-			'address, or constant',
-			'onto the stack.',
-		]);
-	});
-
-	it('generates tooltip color transitions for instructions and stack values', () => {
-		const fontTooltipText = {};
-		const fontTooltipHighlight = {};
-		const content = getSelectedLineTooltipContent(
-			'add',
-			TOOLTIP_WRAP_WIDTH,
-			{
-				lineNumberBeforeMacroExpansion: 2,
-				lineNumberAfterMacroExpansion: 2,
-				instruction: 'add',
-				stackAnalysis: {
-					stackBefore: [
-						{ isInteger: true, knownIntegerValue: 1 },
-						{ isInteger: true, knownIntegerValue: 2 },
-					],
-					consumedOperands: [
-						{ isInteger: true, knownIntegerValue: 1 },
-						{ isInteger: true, knownIntegerValue: 2 },
-					],
-					producedStackItems: [{ isInteger: true }],
-					stackAfter: [{ isInteger: true }],
-				},
-			},
-			{
-				fontTooltipText,
-				fontTooltipHighlight,
-			} as never
-		);
-
-		expect(content.text).toEqual([
-			'add (T T -- T)',
-			'Adds two numbers of the same',
-			'type and pushes the result.',
-			'before [int=1, int=2]',
-			'after: [int]',
-		]);
-		expect(content.characters[0]).toEqual(toCharacters('add (T T -- T)'));
-		expect(content.lineCount).toBe(5);
-		expect(content.widthChars).toBe('Adds two numbers of the same'.length);
-		expect(content.colors[0][0]).toBe(fontTooltipHighlight);
-		expect(content.colors[0][3]).toBeUndefined();
-		expect(content.colors[3][0]).toBe(fontTooltipText);
-		expect(content.colors[3][7]).toBe(fontTooltipHighlight);
-		expect(content.colors[3][12]).toBeUndefined();
-		expect(content.colors[4][0]).toBe(fontTooltipText);
-		expect(content.colors[4][7]).toBe(fontTooltipHighlight);
-	});
-
-	it('formats selected line stack analysis', () => {
-		expect(
-			getStackAnalysisTooltipText({
-				lineNumberBeforeMacroExpansion: 2,
-				lineNumberAfterMacroExpansion: 2,
-				instruction: 'add',
-				stackAnalysis: {
-					stackBefore: [
-						{ isInteger: true, knownIntegerValue: 1 },
-						{ isInteger: true, knownIntegerValue: 2 },
-					],
-					consumedOperands: [
-						{ isInteger: true, knownIntegerValue: 1 },
-						{ isInteger: true, knownIntegerValue: 2 },
-					],
-					producedStackItems: [{ isInteger: true }],
-					stackAfter: [{ isInteger: true }],
-				},
-			})
-		).toEqual(['before [int=1, int=2]', 'after: [int]']);
-	});
-
 	it('adds selected line stack analysis when compiler data is available', () => {
 		const selectedBlock = createMockCodeBlock({
 			code: ['push 1', 'push 2', 'add'],
@@ -189,8 +62,8 @@ describe('tooltip effect', () => {
 			'add (T T -- T)',
 			'Adds two numbers of the same',
 			'type and pushes the result.',
-			'before [int=1, int=2]',
-			'after: [int]',
+			'before [>int=1, >int=2]',
+			'after: [int<]',
 		]);
 	});
