+++ packages/editor/packages/editor-state/src/features/tooltip/sourceLine.ts
@@ -1,0 +1,71 @@
+import {
+	ArgumentType,
+	getInstructionSpec,
+	getInstructionStackSignature,
+	memoryDeclarationInstructions,
+	type AST,
+	type InstructionSpec,
+} from '@8f4e/compiler-spec';
+
+const memoryIdentifierRegExp = /^[a-z_]\w*$/i;
+const memoryDeclarationInstructionSet = new Set<string>(memoryDeclarationInstructions);
+
+export function getInstructionNameFromSourceLine(line: string): string | undefined {
+	return line.trim().split(/\s+/)[0] || undefined;
+}
+
+export function getInstructionSpecFromSourceLine(line: string): InstructionSpec | undefined {
+	const instruction = getInstructionNameFromSourceLine(line);
+
+	if (!instruction || instruction.startsWith(';')) {
+		return undefined;
+	}
+
+	return getInstructionSpec(instruction);
+}
+
+export function getMemoryDeclarationIdFromSourceLine(line: string | undefined): string | undefined {
+	const source = line?.split(';')[0].trim();
+
+	if (!source) {
+		return undefined;
+	}
+
+	const [instruction, id] = source.split(/\s+/);
+
+	if (!memoryDeclarationInstructionSet.has(instruction) || !id || !memoryIdentifierRegExp.test(id)) {
+		return undefined;
+	}
+
+	return id;
+}
+
+function getStackSignatureLineFromSourceLine(line: string, instruction: string): AST[number] | undefined {
+	if (instruction !== 'storeBytes') {
+		return undefined;
+	}
+
+	const count = Number.parseInt(line.trim().split(/\s+/)[1] ?? '', 10);
+
+	if (!Number.isFinite(count)) {
+		return undefined;
+	}
+
+	return {
+		lineNumberBeforeMacroExpansion: 0,
+		lineNumberAfterMacroExpansion: 0,
+		instruction: 'storeBytes',
+		arguments: [{ type: ArgumentType.LITERAL, value: count, isInteger: true }],
+	};
+}
+
+export function getStackSignatureFromSourceLine(line: string): string | undefined {
+	const instruction = getInstructionNameFromSourceLine(line);
+	const spec = getInstructionSpecFromSourceLine(line);
+
+	if (!instruction || !spec) {
+		return undefined;
+	}
+
+	return getInstructionStackSignature(instruction, getStackSignatureLineFromSourceLine(line, instruction));
+}
